# HDDS-12337. Speed up list tests