--- a/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestListKeys.java
+++ b/hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestListKeys.java
@@ -63,12 +63,14 @@ public abstract class TestListKeys implements NonHATests.TestCase {
   private OzoneBucket legacyOzoneBucket;
   private OzoneBucket obsOzoneBucket;
   private OzoneClient client;
-
+  private boolean originalFileSystemPathEnabled;
   private long originalMaxListSize;
 
   @BeforeAll
   void init() throws Exception {
     OmConfig omConfig = cluster().getOzoneManager().getConfig();
+    originalFileSystemPathEnabled = omConfig.isFileSystemPathEnabled();
+    omConfig.setFileSystemPathEnabled(true);
     originalMaxListSize = omConfig.getMaxListSize();
     omConfig.setMaxListSize(2);
 
@@ -93,7 +95,9 @@ public abstract class TestListKeys implements NonHATests.TestCase {
   @AfterAll
   void cleanup() {
     IOUtils.closeQuietly(client);
-    cluster().getOzoneManager().getConfig().setMaxListSize(originalMaxListSize);
+    OmConfig omConfig = cluster().getOzoneManager().getConfig();
+    omConfig.setFileSystemPathEnabled(originalFileSystemPathEnabled);
+    omConfig.setMaxListSize(originalMaxListSize);
   }
 
   private void initFSNameSpace() throws Exception {