## What changes were proposed in this pull request?

Speed up `TestList*` by running with shared cluster.

https://issues.apache.org/jira/browse/HDDS-12337

## How was this patch tested?

Before:

```
[INFO] Tests run: 13, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 37.151 s - in org.apache.hadoop.ozone.om.TestListStatus
[INFO] Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 40.264 s - in org.apache.hadoop.ozone.om.TestListKeysWithFSO
[INFO] Tests run: 24, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 39.002 s - in org.apache.hadoop.ozone.om.TestListKeys
```

After:

```
[INFO] Tests run: 13, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.182 s - in org.apache.ozone.test.NonHATests$ListStatus
[INFO] Tests run: 5, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 3.101 s - in org.apache.ozone.test.NonHATests$ListKeysWithFSO
[INFO] Tests run: 24, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 0.933 s - in org.apache.ozone.test.NonHATests$ListKeys
```

https://github.com/adoroszlai/ozone/actions/runs/13352674274